+++ Changes.md
@@ -1,3 +1,6 @@
+* Fix the syncing of the table names on the score sheet generation page
+* 454 - Add support for multiple playoff brackets on a single big screen display
+
 Release 11.5
 ============